+++ src/scraper/dto/scrape-command.dto.ts
@@ -4,7 +4,6 @@ import { InputTarget } from "../types/input-target.class"
 export class ScrapeCommandDto {
   url: string
   targets?: InputTarget[] = []
-  csr?: boolean = false
   headers?: Record<string, string>
   httpProxy?: HttpProxy
 }